Add delete button on pages list
FossilOrigin-Name: 42f67999dd4c00090669c6f9351ad227cced0299f4319808ba2f3130fb9b9f67

+++ src/include/lib/Paheko/Entities/Web/Page.php
@@ -206,12 +206,14 @@ class Page extends Entity
 	{
 		$dir = $this->dir();
 
-		if ($dir) {
+		$r = parent::delete();
+
+		if ($r && $dir) {
 			$dir->delete();
 		}
 
 		Cache::clear();
-		return parent::delete();
+		return $r;
 	}
 
 	public function selfCheck(): void
